LayoutTests/http/wpt/webcodecs/videoFrame-colorSpace.html

<!DOCTYPE html>
<html>
<header>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
</header>
<body>
<script>

function makeOffscreenCanvas(width, height) {
  let canvas = new OffscreenCanvas(width, height);
  let ctx = canvas.getContext('2d');
  ctx.fillStyle = 'rgba(50, 100, 150, 255)';
  ctx.fillRect(0, 0, width, height);
  return new VideoFrame(canvas, { timestamp: 1 });
}

async function doEncodeDecode(encoderConfig)
{
  let resolve, reject;
  const promise = new Promise((res, rej) => {
    resolve = res;
    reject = rej;
  });

  const decoder = new VideoDecoder({
    output(frame) {
      resolve(frame);
    },
    error(e) {
      reject(e.message);
    }
  });

  const encoderInit = {
    output(chunk, metadata) {
      let config = metadata.decoderConfig;
      if (config) {
        decoder.configure(config);
      }
      decoder.decode(chunk);
    },
    error(e) {
      reject(e.message);
    }
  };

  const encoder = new VideoEncoder(encoderInit);
  encoder.configure(encoderConfig);

  const w = encoderConfig.width;
  const h = encoderConfig.height;
  const frame = makeOffscreenCanvas(w, h);
  encoder.encode(frame, { keyFrame: true });

  await encoder.flush();
  await decoder.flush();
  encoder.close();
  decoder.close();

  return promise;
}

function doTest(codec, title)
{
  const config = { codec };
  config.width = 320;
  config.height = 200;
  config.bitrate = 1000000;
  config.framerate = 30;

  promise_test(async t => {
    const frame = await doEncodeDecode(config);
    t.add_cleanup(() => frame.close());
 
    assert_not_equals(frame.colorSpace.primaries, null, "primaries");
  }, title);
}

doTest('vp8', "VP8 decoded frame color space");
doTest('vp09.00.10.08', "VP9 decoded frame color space");
doTest('avc1.42001E', "H.264 decoded frame color space");
</script>
</body>
</html>
